Try my 1.5.6 stuff which uses native code
This was the stuff I was trying out migrating arbotix to 1.6 release of
Arduino.
WIP WIP WIP

Unlike other releases this code does not replicate the Arduino code
base, it simply uses it, but defines a new board type for Arbotix.  I
included a version of the Bioloid library (with the name
BioloidSerial) that uses the actual Serial class.  It worked for me on
earlier builds of Arduino, have not tested yet...